+++ packages/lambda/src/cli/commands/regions.ts
@@ -1,0 +1,8 @@
+import {Log} from '@remotion/cli/dist/log';
+import {getRegions} from '../..';
+
+export const REGIONS_COMMAND = 'regions';
+
+export const regionsCommand = () => {
+	Log.info(getRegions().join(' '));
+};