+++ packages/lambda/src/cli/help.ts
@@ -1,9 +1,11 @@
 import {CliInternals} from '@remotion/cli';
 import {BINARY_NAME} from '../shared/constants';
 import {FUNCTIONS_COMMAND} from './commands/functions';
+import {POLICIES_COMMAND} from './commands/policies/policies';
+import {REGIONS_COMMAND} from './commands/regions';
 import {RENDER_COMMAND} from './commands/render/render';
-import {STILL_COMMAND} from './commands/still';
 import {SITES_COMMAND} from './commands/sites';
+import {STILL_COMMAND} from './commands/still';
 import {Log} from './log';
 
 const packagejson = require('../../package.json');
@@ -31,4 +33,12 @@ export const printHelp = () => {
 	Log.info();
 	Log.info(`${BINARY_NAME} ${SITES_COMMAND}`);
 	Log.info(CliInternals.chalk.gray('Deploy and manage Remotion projects.'));
+
+	Log.info();
+	Log.info(`${BINARY_NAME} ${POLICIES_COMMAND}`);
+	Log.info(CliInternals.chalk.gray('View and validate AWS policy files.'));
+
+	Log.info();
+	Log.info(`${BINARY_NAME} ${REGIONS_COMMAND}`);
+	Log.info(CliInternals.chalk.gray('Show the list of AWS regions supported.'));
 };